pyproject.toml
[project]
name = "cvmatch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "langchain>=1.2.9",
    "langchain-anthropic>=1.3.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "traceback-with-variables>=2.2.1",
]
